refactored flight traits and un-inheritance flight subclasses App\Models\Flights\MasterFlight/Flight classes now deprecated
